requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
pandas==1.4.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.9
